moved util functions back as static functions to Form
